Move types and values for options in own file
