+++ lib/node/sources/BufferSource.ts
@@ -1,4 +1,4 @@
-import type { FileSource } from '../../upload.js'
+import type { FileSource } from '../../options.js'
 
 export default class BufferSource implements FileSource<Buffer> {
   size: number
