+++ lib/node/sources/StreamSource.ts
@@ -1,5 +1,5 @@
 import type { Readable } from 'node:stream'
-import type { FileSource } from '../../upload.js'
+import type { FileSource } from '../../options.js'
 
 /**
  * readChunk reads a chunk with the given size from the given
